refactor hint to separate component
